#5 [fix] 에러 핸들링 관련 수정
